New lock mechanishm for XIA routing tables
The previous lock mechanishm simply relied on RTNL lock,
whereas the new one uses:
1. RCU for reads,
2. RTNL lock when RTNL is the only writer,
3. A readers-write spin lock and a table of spin locks for writers.

Although the internals of the new routing tables is rather different for
single writer and multiple writers the interface is the same,
the difference is restricted to the initialization.

Added reference counting to struct fib_xid_table (xtbl) to allow
external references to it.

The new code has been minimally tested.