Fix neg. dependency bug related to XID redirects
When a known principal redirects to unknown XID, being the latter
of a known or unknown principal, and the negative anchor is
a single anchor to limit cost while routing, one can't just
filter by the unkown XID in the negative anchor because
it won't find all XDST entries that must be flushed.
The correct solution in this case is to flush
the whole negative anchor.

If a principal finds necessary to not flush the whole negative
anchor every time, intead of paying the full price of keeping
negative entries in its XID table, the principal can implement
a hash of negative anchor to split the single negative anchor.
The key of the hash would be the unknown XID.