Define and alter functions for sending monitoring information
Because neighbors now have a status associated with them, NWP
needs methods of disseminating that information. To do this,
the routines for announcing and sending neighbor lists must
be amended to include status information.

send_monitoring() is also defines for convenience in sending
the different kinds of monitoring packets. If the investigator
is not needed, then the const u8 *inv argument should be NULL.