Enable receipt and processing of monitoring packets
nwp_rcv() may now accept monitoring packets and distribute
them to process_monitoring(), which updates clocks, changes
timers, and dispatches packets as necessary depending on the
type of packet that has been received.

Note that the reserved field of monitoring packets MUST be
non-zero as a convention for acceptance by nwp_rcv().

Since monitoring packets may now be received, it is possible
for monitoring to be successful (the target responds). Hence,
end_interval_success() is added.