Add functionality for maintaining neighbor status
struct hrdw_addr needs fields describing the local clock and remote
status (failure status and clock time) of a neighbor. The first 31
bits of remote_sc are the remote node's clock, and the 32nd bit of
remote_rc is the failure status (0 for failed, 1 for alive).
local_c is 32 bits of the local clock.

All clock times are taken from the lower 32 bits of the wall time,
measured in seconds. When two remote clock values need to be
directly compared (only the lower 31 bits represent the clock), the
function clockcmp31() should be used, which is a wrapper for
clockcmp32(). When two local clocks need to be compared (where all
32 bits represent the clock), clockcmp32() can be directly called.
These functions correctly handle overflow of clock values.

Because neighbors now have a status associated with them, NWP
needs methods of disseminating that information. To do this,
the routines for announcing and sending neighbor lists must
be amended to include status information, and the packet definitions
for announcements and neighbor lists must be amended. A node will
only send information about a failed neighbor if it is within a
time bound defined by NWP_LIST_TIME_MAX so that failed neighbors
are not cyclically added and deleted from neighbor lists in the LAN.

Status information can be taken from announcement and neighbor list
packets in read_announcement() and process_neigh_list(), and can be
passed to functions that create and insert neighbors, such as
new_ha(), add_ha(), and insert_neigh().

main_newroute() in net/xia/ppal_hid/main.c must also be altered
to match the change in function signature for insert_neigh(),
and the function prototype must be changed in
include/net/xia_hid.h.