net: run GC only when new dst entry is accounted
Without this patch, GC of DST table run even for entries that are
not supposed to be accounted.

Given that this is a change in the core of Linux's network
infrastruture, it should be discussed with maintainers to add, or
not, it into mainline.

XIP needs it because flag DST_NOCOUNT is used for dst entries
that have a very short lifespan, and as such should not
influence GC.

Once dependency forest is fully implemented, this patch won't be
necessary anymore, and can be removed.

REMOVE ME WHEN POSSIBLE!

Signed-off-by: Michel Machado <michel@digirati.com.br>