XIA/U4ID: add u4id_deliver and DST I/O functions
If the U4ID to be routed is a local U4ID and it is not a sink, we
have reached the destination of an XIP tunnel. Therefore, the last
node should be advanced and routing may continue.

If the U4ID to be routed has a valid XID but is not local, we
need to tunnel.