XIA/U4ID: add local U4IDs
Routines for creating, deleting, dumping, and freeing local U4IDs are
added. When a local U4ID is added, a socket is opened on the given
address:port. The socket is marked as an encapsulation socket, which
allows a callback function, u4id_udp_encap_recv, to be registered, which
will be called upon packet delivery on that socket.

Upon receiving an encapsulated XIP packet, u4id_udp_encap_recv
puts the packet on the loopback device so that the packet is
naturally delivered to the xip_rcv function. This passes control
of the packet from the IP stack to the XIA stack.

A new UDP encapsulation type is added to include/uapi/linux/udp.h.

Since local_free_u4id is invoked in an atomic context, the socket
cannot be destroyed. Therefore, a workqueue is used to defer the
socket destruction (and the freeing of the struct fib_xid_u4id_local).