XIA/U4ID: improve local_delroute()
local_delroute() was assuming that the caller held an RCU read lock.
That was true, but there are no explicit guarantee that will always
be the case.

This patch also deals with a race condition that can take place
in applications that quickly edits the routing table;
see code for details.

Finally, this patch clarifies comments around
the activation/deactivation of the tunnel socket.