xia/serval: prefix sk_stream_alloc_skb() with `serval_'
While compiling Serval as part of the kernel (i.e. not as a module),
sk_stream_alloc_skb() conflicts with
the same-name function of TCP.

The problem does not come up when compiling Serval as a module
because TCP does not export that function;
although sk_stream_alloc_skb() is not static either.

Reported-by: Alexander Breen <alexanderbreen@gmail.com>
Reported-by: Yuguang Li <leeygxz@gmail.com>
Signed-off-by: Michel Machado <michel@digirati.com.br>