xia/serval: deal with dropped `when' field
Field `when' was dropped from struct tcp_skb_cb in
the following commit:
7faee5c0d514162853a343d93e4a0b6bb8bfec21
tcp: remove TCP_SKB_CB(skb)->when

This patch add field `when' for Serval defining
struct serval_tcp_skb_cb.

This solution is not ideal, but it buys us time.
The long-term solution is to make TCP support working on top of
any network protocol, and bring TCP to XIA as a principal.