xia: add ``list'' description to FIB functions
The FIB code in XIA assumes that the underlying data structure
is a list of buckets. To separate the list-specific FIB code from
the general FIB code, any function in net/xia/fib.c (as well as
include/net/xia_fib.h) that is exported and uses buckets was
renamed to include the ``list'' descriptor. Additionally,
functions in net/xia/fib.c that have generic names, even if they
are static, were renamed (such as find_xid_locked -->
find_list_xid_locked) if they use buckets. This cleans up the
namespace for future changes.

The one exception to this patch is the function xtbl_put().
Although xtbl_put() is technically list FIB centric in this
patch, it will eventually be made generic, and is best left
alone.