xia: add fxid_copy to FIB interface
Since the underlying structure of a FIB entry is not known to
the principal, it cannot copy a FIB entry. This patch adds a
function to every FIB that makes a copy of a FIB entry that
includes FIB-specific data.

This patch also fixes a bug in the LPM principal by making use
of the new copy function.